# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DeleteResponse {
  success: boolean;
}

export type Direction = (typeof Direction)[keyof typeof Direction];

export const Direction = {
  Long: "Long",
  Short: "Short",
} as const;

export type SetupType = (typeof SetupType)[keyof typeof SetupType];

export const SetupType = {
  Pullback: "Pullback",
  Breakout: "Breakout",
  Reversal: "Reversal",
} as const;

export type Session = (typeof Session)[keyof typeof Session];

export const Session = {
  London: "London",
  NY: "NY",
  Asia: "Asia",
} as const;

export type ExecutionQuality =
  (typeof ExecutionQuality)[keyof typeof ExecutionQuality];

export const ExecutionQuality = {
  "A+": "A+",
  B: "B",
  FOMO: "FOMO",
} as const;

export interface Trade {
  id: string;
  symbol: string;
  entryPrice: number;
  exitPrice: number;
  size: number;
  direction: Direction;
  tradedAt: string;
  setupType: SetupType;
  session: Session;
  emaAlignment: boolean;
  executionQuality: ExecutionQuality;
  notes?: string | null;
  screenshotUrl?: string | null;
  pnl: number;
  rr: number;
  createdAt: string;
}

export interface TradeInput {
  /** @minLength 1 */
  symbol: string;
  /** @exclusiveMinimum 0 */
  entryPrice: number;
  /** @exclusiveMinimum 0 */
  exitPrice: number;
  /** @exclusiveMinimum 0 */
  size: number;
  direction: Direction;
  tradedAt: string;
  setupType: SetupType;
  session: Session;
  emaAlignment: boolean;
  executionQuality: ExecutionQuality;
  notes?: string | null;
  screenshotUrl?: string | null;
}

export interface MetricsSummary {
  totalPnl: number;
  tradeCount: number;
  winRate: number;
  avgWin: number;
  avgLoss: number;
  expectancyR: number;
  wins: number;
  losses: number;
  breakeven: number;
}

export interface EquityPoint {
  date: string;
  cumulativePnl: number;
}

export interface WinLoss {
  wins: number;
  losses: number;
  breakeven: number;
}

export interface GroupedPerformance {
  label: string;
  tradeCount: number;
  winRate: number;
  totalPnl: number;
  avgPnl: number;
}

export interface ExecutionComparison {
  groupA: string;
  groupB: string;
  winRateA: number;
  winRateB: number;
  pnlA: number;
  pnlB: number;
  countA: number;
  countB: number;
}

export interface ExecutionAnalysis {
  emaComparison: ExecutionComparison;
  qualityComparison: ExecutionComparison;
}

export interface WeeklyReview {
  weekStart: string;
  weekEnd: string;
  tradeCount: number;
  totalPnl: number;
  bestSetup?: string | null;
  worstSetup?: string | null;
  bestSession?: string | null;
  mostCommonMistake?: string | null;
}

export type ListTradesParams = {
  setup?: SetupType;
  session?: Session;
  execution?: ExecutionQuality;
  sortBy?: ListTradesSortBy;
  sortDir?: ListTradesSortDir;
};

export type ListTradesSortBy =
  (typeof ListTradesSortBy)[keyof typeof ListTradesSortBy];

export const ListTradesSortBy = {
  date: "date",
  pnl: "pnl",
} as const;

export type ListTradesSortDir =
  (typeof ListTradesSortDir)[keyof typeof ListTradesSortDir];

export const ListTradesSortDir = {
  asc: "asc",
  desc: "desc",
} as const;