# lib/api-zod/src/generated/types/direction.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export type Direction = (typeof Direction)[keyof typeof Direction];

export const Direction = {
  Long: "Long",
  Short: "Short",
} as const;