# lib/api-zod/src/generated/types/executionAnalysis.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionComparison } from "./executionComparison";

export interface ExecutionAnalysis {
  emaComparison: ExecutionComparison;
  qualityComparison: ExecutionComparison;
}