# lib/api-zod/src/generated/types/executionQuality.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export type ExecutionQuality =
  (typeof ExecutionQuality)[keyof typeof ExecutionQuality];

export const ExecutionQuality = {
  "A+": "A+",
  B: "B",
  FOMO: "FOMO",
} as const;