# lib/api-zod/src/generated/types/groupedPerformance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface GroupedPerformance {
  label: string;
  tradeCount: number;
  winRate: number;
  totalPnl: number;
  avgPnl: number;
}