# lib/api-zod/src/generated/types/metricsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface MetricsSummary {
  totalPnl: number;
  tradeCount: number;
  winRate: number;
  avgWin: number;
  avgLoss: number;
  expectancyR: number;
  wins: number;
  losses: number;
  breakeven: number;
}