# lib/api-zod/src/generated/types/setupType.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export type SetupType = (typeof SetupType)[keyof typeof SetupType];

export const SetupType = {
  Pullback: "Pullback",
  Breakout: "Breakout",
  Reversal: "Reversal",
} as const;