# lib/api-zod/src/generated/types/trade.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { Direction } from "./direction";
import type { ExecutionQuality } from "./executionQuality";
import type { Session } from "./session";
import type { SetupType } from "./setupType";

export interface Trade {
  id: string;
  symbol: string;
  entryPrice: number;
  exitPrice: number;
  size: number;
  direction: Direction;
  tradedAt: Date;
  setupType: SetupType;
  session: Session;
  emaAlignment: boolean;
  executionQuality: ExecutionQuality;
  notes?: string | null;
  screenshotUrl?: string | null;
  pnl: number;
  rr: number;
  createdAt: Date;
}