# lib/api-zod/src/generated/types/weeklyReview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyReview {
  weekStart: Date;
  weekEnd: Date;
  tradeCount: number;
  totalPnl: number;
  bestSetup?: string | null;
  worstSetup?: string | null;
  bestSession?: string | null;
  mostCommonMistake?: string | null;
}