# lib/api-zod/src/generated/types/winLoss.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TradeTrack Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface WinLoss {
  wins: number;
  losses: number;
  breakeven: number;
}